feat: added bids and rentalpost to context provider
