+++ src/features/escrow/components/escrow-lifecycle-header.tsx
@@ -1,6 +1,6 @@
 import { Card, CardContent } from '@/components/ui/card';
 import { Progress } from '@/components/ui/progress';
-import { useEscrowLifecycle } from './escrow-lifecycle-context';
+import { useEscrowLifecycle } from '../contexts/escrow-lifecycle-context';
 import { Doc } from '@convex/_generated/dataModel';
 
 interface EscrowLifecycleHeaderProps {
