+++ src/features/escrow/components/escrow-step.tsx
@@ -10,7 +10,7 @@ import { Step5Pending } from './StepContentViews/step-5-pending';
 import { TransactionDetails } from './transaction-details';
 import { Doc } from '@convex/_generated/dataModel';
 import { StepStatus } from '../types/step-status';
-import { useEscrowLifecycle } from './escrow-lifecycle-context';
+import { useEscrowLifecycle } from '../contexts/escrow-lifecycle-context';
 
 interface EscrowStepProps {
   step: Doc<'escrowSmartContractSteps'>;
@@ -44,13 +44,12 @@ export function EscrowStep({ step, index }: EscrowStepProps) {
       animate={{ opacity: 1, y: 0 }}
       transition={{ duration: 0.5, delay: index * 0.1 }}>
       <Card
-        className={`border-slate-800 ${
-          step.status === 'COMPLETED'
+        className={`border-slate-800 ${step.status === 'COMPLETED'
             ? 'bg-green-900/20 border-green-800'
             : step.status === 'ACTIVE'
               ? 'bg-blue-900/20 border-blue-800'
               : 'bg-slate-900/50'
-        }`}>
+          }`}>
         <CardContent className='p-6'>
           <div className='flex items-start space-x-4'>
             <div className='flex-shrink-0 mt-1'>{getStepIcon(step.status as StepStatus)}</div>
