+++ src/features/escrow/contexts/escrow-lifecycle-context.tsx
@@ -2,6 +2,7 @@ import { createContext, useContext, ReactNode } from 'react';
 import { Doc } from '@convex/_generated/dataModel';
 
 interface EscrowLifecycleContextProps {
+  bid: Doc<'bids'>;
   rentalPost: Doc<'rentalposts'>;
   escrowData: Doc<'escrowSmartContracts'>;
   timeRemaining: {
@@ -14,14 +15,17 @@ const EscrowLifecycleContext = createContext<EscrowLifecycleContextProps | undef
 
 export const useEscrowLifecycle = () => {
   const context = useContext(EscrowLifecycleContext);
+
   if (!context) {
     throw new Error('useEscrowLifecycle must be used within an EscrowLifecycleProvider');
   }
+
   return context;
 };
 
 interface EscrowLifecycleProviderProps {
   children: ReactNode;
+  bid: Doc<'bids'>;
   rentalPost: Doc<'rentalposts'>;
   escrowData: Doc<'escrowSmartContracts'>;
   timeRemaining: {
@@ -30,8 +34,16 @@ interface EscrowLifecycleProviderProps {
   };
 }
 
-export const EscrowLifecycleProvider = ({ children, rentalPost, escrowData, timeRemaining }: EscrowLifecycleProviderProps) => {
+export const EscrowLifecycleProvider = ({
+  children,
+  bid,
+  rentalPost,
+  escrowData,
+  timeRemaining,
+}: EscrowLifecycleProviderProps) => {
   return (
-    <EscrowLifecycleContext.Provider value={{ rentalPost, escrowData, timeRemaining }}>{children}</EscrowLifecycleContext.Provider>
+    <EscrowLifecycleContext.Provider value={{ bid, rentalPost, escrowData, timeRemaining }}>
+      {children}
+    </EscrowLifecycleContext.Provider>
   );
 };
